When setting -fsanitize=undefined also set it as linker flag
I'm actually not really happy that we get this into our builds,
at least it's only debug. But if this is going to be set, then
set it for real, so that linking with clang on Linux also works.

This helps me running clazy and clang-tidy which are easier to do with
clang on Linux but don't link without this. It's actually surprising
that it works on macOS and with GCC.