Fix building as dynamic lib
Let's not make things complicated. As of now, qtwebapp only supports
building as dynamic library, even on Windows. Sadly clang was forgotten,
so instead of having a lot of code trying to get import/export right,
build on top of Qt's macros.
And use the auto generated qtwebappglobal.h.in instead of having a
broken copy ourselves.